fix(api): close the integration gaps the interrupted workstreams left
mypy strict found real breaks at the seams, not just style:

- The worker called ingest_source/render_sheet_pdfs with a Job where a UUID was
  expected, and called scan_service.process_scan and adaptive_service.
  generate_adaptive, neither of which exists. Every background job would have
  failed at runtime.
- Nothing implemented PROCESS_SCAN at all: create_scan queued a job that no code
  consumed, so the F2 loop was broken end to end. alppy/services/
  scan_processing.py now decodes the upload, registers each page, reads the UID
  and persists detections.
- That exposed a subtlety worth naming: on a differentiated sheet each student's
  copy has a different item list, so "how many bubbles are on row 3" differs per
  copy. Detection is two passes -- read the UID against a full grid, then
  re-detect against that student's own printed layout -- rather than assuming
  one copy's geometry for the whole pile.
- A missing "from typing import Final" in deps.py, masked only because
  "from __future__ import annotations" defers annotation evaluation.
- Two loops in the ingest pipeline reused the name "old" for a SourceChunk and
  then for an Exercise.

Adds run_seed: the demo school, class 7B, 18 invented students and three weeks
of simulated lessons, idempotent on every boot. The band distribution comes out
of the real mastery model rather than being written down. Recalibrated the
ability profiles after the first run produced a class that was 44% "fading" --
a median student reading as failing on their own textbook exercises is a broken
demo, not a candid one.

ruff clean, mypy strict clean across 65 files, 277 tests pass.

+++ apps/api/alppy/scan/synthetic.py
@@ -38,7 +38,7 @@ class SyntheticSheet:
 
 
 def _mm(v: float) -> int:
-    return int(round(v * PX_PER_MM))
+    return round(v * PX_PER_MM)
 
 
 def render_page(
@@ -88,7 +88,7 @@ def render_page(
         for oi in range(n_options):
             cx, cy = L.bubble_centre_mm(item_index, oi)
             centre = (_mm(cx), _mm(cy))
-            radius = int(round(L.BUBBLE_D_MM / 2.0 * PX_PER_MM))
+            radius = round(L.BUBBLE_D_MM / 2.0 * PX_PER_MM)
             cv2.circle(img, centre, radius, INK, thickness=1)
             if marked[item_index] == oi:
                 shade = int(PAPER * (1.0 - pencil))
